+++ lib/templates/with-express/src/custom/index.ts
@@ -1,0 +1,12 @@
+import { Request } from './express.request'
+import { Response } from './express.response'
+import { CustomError, ICustomError } from './error'
+import { CustomNodeJSGlobal } from './global.variables'
+
+export {
+  CustomError,
+  CustomNodeJSGlobal,
+  ICustomError,
+  Response,
+  Request
+}