+++ lib/templates/with-express/src/dto-interfaces/index.ts
@@ -1,0 +1,3 @@
+import { DtoUser } from './users.dto'
+
+export { DtoUser }