+++ lib/templates/with-express/src/dto-interfaces/users.dto.ts
@@ -1,5 +1,7 @@
-export interface DtoUsers {
+interface DtoUser {
   id       : string,
   lastName?: string,
   name    ?: string,
 }
+
+export { DtoUser }