+++ lib/templates/with-express/src/utils/index.ts
@@ -1,0 +1,3 @@
+import { response } from './response'
+
+export { response }